python-iam-ai/requirements.txt
boto3
mysql-connector-python
openai  # Optional AI integration
